+++ src/Animations/index.ts
@@ -1,1 +1,2 @@
 export {default as LaptopAnimation} from "./laptop-animation.json";
+export {default as GameMachine} from "./game-machine.json";